### 2021-02-18 #104 @whedon
[PRE REVIEW]: Auto-eD: A visual learning tool for automatic differentiation
**Submitting author:** @lindseysbrown (<a href="http://orcid.org/0000-0002-9387-1387">Lindsey Brown</a>)
**Repository:** <a href="https://github.com/lindseysbrown/Auto-eD" target ="_blank">https://github.com/lindseysbrown/Auto-eD</a>
**Version:** v1.0.0

@lindseysbrown if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
