### 2021-06-28 #104 @labarba
hi @Atcold 👋  —  thanks for agreeing to be a review for JOSE! I'll be looking for a second reviewer, and when added here, will start another GitHub where the review happens. In the meantime, have a look at our [documentation](https://openjournals.readthedocs.io/) and ask me anything!
