### 2021-06-28 #104 @labarba
@amueller — We're in need of a second reviewer for this submission to the Journal of Open Source Education, JOSE. Would you be amenable to contributing a review?
"Auto-eD: A visual learning tool for automatic differentiation"
[View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00104/jose.00104/10.21105.jose.00104.pdf)
