### 2021-07-03 #104 @labarba
I'm still looking for a second reviewer for this submission. @JontySinai — could I interest you in contributing a review for the Journal of Open Source Education, [JOSE](https://jose.theoj.org)? The submission is:
**"Auto-eD: A visual learning tool for automatic differentiation"**
[View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00104/jose.00104/10.21105.jose.00104.pdf)
