### 2021-07-13 #104 @labarba
@rasbt – Would you be willing to contribute a review on this submission to JOSE? I know you recently reviewed for us, and I don't want to lean to heavily on you, but if you can find the cycles, I'm needing the help. If you cannot do it, but can offer some suggested names for alternative reviewers, I'd be grateful.
Submission title: **"Auto-eD: A visual learning tool for automatic differentiation"**
[View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00104/jose.00104/10.21105.jose.00104.pdf)
