### 2021-07-14 #104 @rasbt
@labarba Thanks for considering me, but I am very time-constrained in the upcoming 1 1/2 months since I will be traveling and have limited access to my computer.

@Atcold might possibly willing (and possibly excited) to review this article :)
