### 2021-07-15 #104 @rasbt
Oops, I should have double-checked :). Below is a list of machine learning researchers and educators who would be great candidates:

@brohrer
@khakhalin
@muellerzr
@radekosmulski
