### 2021-07-25 #104 @labarba
@brohrer Hi Brandon 👋  — I see on your site that you have a lot of educational material. Could I interest you in contributing a review for this submission to JOSE, the Journal of Open Source Education?

> Auto-eD: A visual learning tool for automatic differentiation
> [View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00104/jose.00104/10.21105.jose.00104.pdf)

[JOSE](https://jose.theoj.org) is a community led open-source journal to publish articles about educational software and learning modules (human learning, not machine, ha ha).

I already have one reviewer for this one, but need a second one to initiate the review. Have a look and let me know!
