### 2021-07-27 #104 @brohrer
Thanks for thinking for me on this, it’s a big vote of confidence. I’m afraid I will have to decline. I have fully booked myself for the rest of the year and I’m having to be extra disciplined in my prioritization.

Sorry!
Brandon
