### 2021-07-28 #104 @labarba
Thanks, @brohrer — another time! (Feel free to unsubscribes to this thread to avoid more notifications.)

@khakhalin Hi Arseny 👋  — would you consider contributing a peer review for JOSE? (The Journal of Open Source Education)
Submission title: **"Auto-eD: A visual learning tool for automatic differentiation"**
[View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00104/jose.00104/10.21105.jose.00104.pdf)
