### 2021-10-13 #104 @khakhalin
Hi Lorena! I'm super-sorry, but I'm still totally under with paperwork and
all the new responsibilities. And also the paper is still not at all close
to anything I ever did before. I'd had to learn a new field to review it,
or at least build some basic idea about it, and I simply don't have time
for this now, even though it is of course very interesting. I'm really
sorry, but I don't think I can help here!

On Thu, Oct 7, 2021 at 4:48 PM Lorena A. Barba ***@***.***>
wrote:

> hey @khakhalin <https://github.com/khakhalin> — We are still unable to
> find a second reviewer for this submission to the Journal of Open Source
> Education, title: *"Auto-eD: A visual learning tool for automatic
> differentiation"*. Would you mind taking a look? It may be a lighter task
> than you thought 2 months ago? Here is the URL of the web app in question:
> https://autoed.herokuapp.com ... and here is the review criteria for
> JOSE: https://openjournals.readthedocs.io/en/jose/review_criteria.html
> Let me know if you could help us out here. We are a bit stuck!
>
