### 2021-11-02 #104 @labarba
hi @merveenoyan — Would you consider volunteering a review for JOSE? (The Journal of Open Source Education, sibling journal to JOSS). I already have one reviewer, but need a second to start the review.
Submission title: **"Auto-eD: A visual learning tool for automatic differentiation"**
[View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00104/jose.00104/10.21105.jose.00104.pdf)
Have a look and let me know! 🙏
