### 2021-12-09 #104 @labarba
hi @rasbt — We're continuing our search for a second reviewer for this JOSE submission, so far with no luck 😞
Do you think you might be able to help us out, now that the semester is closing? Thanks for considering it!
