### 2021-12-28 #104 @labarba
@Atcold 👋 — You had [agreed to review](https://github.com/openjournals/jose-reviews/issues/104#issuecomment-870009251) this submission for us a few months ago. I had a very hard time identifying a second reviewer, and so a lot of time has passed. I want to double-check after all this time that you are still willing and able to review.
Title: **Auto-eD: A visual learning tool for automatic differentiation**
[View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00104/jose.00104/10.21105.jose.00104.pdf)
