### 2022-02-24 #104 @labarba
In all my editing experience, it has _never_ been so hard to secure reviewers! Even though we are a grass-roots journal with no parasitic publisher raking in the big bucks, and folks are generally more willing to review, it is a struggle. I will give it another push!

Thank you @rasbt for your willingness to help. Yes, you are right that we need to start the review (in a new issue) with the checklist.
