### 2022-02-25 #161 @whedon
[REVIEW]: Auto-eD: A visual learning tool for automatic differentiation
**Submitting author:** @lindseysbrown (<a href="http://orcid.org/0000-0002-9387-1387">Lindsey Brown</a>)
**Repository:** <a href="https://github.com/lindseysbrown/Auto-eD" target ="_blank">https://github.com/lindseysbrown/Auto-eD</a>
**Version:** v1.0.2

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).

- [x] **Version**: Does the release version given match the GitHub release (v1.0.0)? [[see external #27](https://github.com/lindseysbrown/Auto-eD/issues/27)]

- [x] **Installation:** Does installation proceed as outlined in the documentation? (and documentation is sufficient?) [[see external #28](https://github.com/lindseysbrown/Auto-eD/issues/28)]

- [x] **Tests:** Are there automated tests or manual steps described so that the function of the software can be verified? [[see external #32](https://github.com/lindseysbrown/Auto-eD/issues/32)]
- [x] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support  [[see external #34](https://github.com/lindseysbrown/Auto-eD/issues/34)]

- [x] **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)? [[see external #35](https://github.com/lindseysbrown/Auto-eD/issues/35)]

- [x] #164
