### 2022-02-25 #161 @labarba
@rasbt, @Atcold — Thank you for agreeing to review for JOSS! This is where the action happens: work your way through the review checklist, feel free to ask questions or post comments here, and also open issues in the submission repository as needed. Godspeed! 🚀
