### 2022-02-25 lindseysbrown/Auto-eD#27 @rasbt
Version number
Hey there,
I am currently reviewing this submission for JOSE and was going through the checklist. One item is

> Version: Does the release version given match the GitHub release (v1.0.0)?

Could you add the version number to the code (maybe I am not seeing it) and make a GitHub release with that version?
