### 2022-03-02 #161 @Atcold
# Drafting here my review

It's *really hard* to understand what the repo / project offers.
There seem to be two components: a Python library and a web app.
The web app is hard to use without a tutorial of sorts. The documentation seems more like a lecture, and only later there is a demo of how to use the web app.

The `README.md` or the PDF paper (or both) should clearly state how the project is organised and how one should consume the material. Right now, I count 3 places where to read (paper, developer manual, documentation) and 2 places where to use code (web app, Python library).

If the project is about providing a tool for understanding automatic differentiation, then the repo should display a tutorial as the main component. Or, even better, the web app should walk you through a teaching experience, but I understand this would be harder to put together.
