### 2022-03-02 #161 @rasbt
@labarba General question about opening issues during the review. Should we be using the "convert to Issue" function here and tagging the author in that issue?
![Screen Shot 2022-03-02 at 3 41 08 PM](https://user-images.githubusercontent.com/5618407/156454318-67c8f6c6-7e5d-4d9a-a8a0-9a1bd20a73dd.png)

Or should we be creating the issues in the project's GitHub repo? Just wondering what the best practices are for JOSE.
