### 2022-03-02 #161 @rasbt
@Atcold I have a similar experience.

Looking at the GitHub repository, I think most of the files (except ./docs) are the ones that are uploaded to Heroku for creating the website? Ideally, it should be more clear what we are looking at in this repo. This could be addressed by clarifying the code structure in the Readme file.

@Atcold One of us should probably create an issue for that. Since you commented first and said it better, do you want to go ahead with that?

The other issue is that it's not quite clear what we are reviewing here. Maybe @labarba could also help clarifying this?

- Should we be reviewing the whole automatic differentiation lecture or just the components that use the web app?
- Should be reviewing the Python & webapp code at all, or should we just focus on the web app as the main "product"?
