### 2022-03-04 #161 @labarba
The author submitted this under the "software" article type, not the "learning module," so I suppose the author intends this application to be used as software to support teaching and learning in machine learning courses. If the author does not make clear how they intend their contribution to be re-used, that is useful input for improvement.
