### 2022-03-04 #161 @lindseysbrown
@labarba Thanks for the clarification.  Our submission spanned both so we weren't sure how to submit.

The main contribution is the web app to visualize automatic differentiation that could be used to support any class in which that is taught.  We have additionally provided a learning module that is complemented by the use of the web app as an example of how the app could be incorporated.  For advanced users with programming experience, we have also provided the underlying software package which can be run independently of the web app.

We can certainly edit the README/paper to make it more clear that you can pick which of these tools are most appropriate based on your level of familiarity with automatic differentiation and with programming:

- Complete beginners should use the Read the Docs for an introduction to automatic differentiation with examples and exercises, complemented by the web app.
- People familiar with automatic differentiation who want a tool to support their understanding should use the web app.
- Advanced programmers who want to explore automatic differentiation in code may use the full software package.
