### 2022-03-21 lindseysbrown/Auto-eD#27 @lindseysbrown
I have now added:
1. Version badge to the readme.
2. Tagged the git repo with the version number, so that git describe command works.
3. Added version.py file
