### 2022-03-21 #161 @lindseysbrown
@Atcold I think you said you were planning on opening an issue, but I didn't see it in the target repo.  I have now added text to both the README and paper (second paragraph in Content) detailing the three different levels of interaction with the software we've provided.

@rasbt @labarba For your reference that this change has been made.  (Still not completely clear on if this was more appropriately classified as software or learning module since we did both.)
