### 2022-03-21 #161 @rasbt
Thanks for making those updates @lindseysbrown !!

Regarding the scope of the review. I am still unsure what to do here and would appreciate your feedback and clarification @labarba . I recently read through the learning modules and had some thoughts, but I am not sure if this is appropriate/within the scope of this review. When I am looking through the reviewer checklist at the top of this thread, the review is basically just about making sure the software works (i.e., installs correctly and works as promised). This is very different from a regular journal review, and I am happy to stick to that, but I want to make sure that this is indeed what you want us to do?
