### 2022-03-21 lindseysbrown/Auto-eD#27 @rasbt
Awesome, thanks! I can see the version in the Readme now, and there is also the version.py file.

Sorry, but I still can't see the tag though. When I go to https://github.com/lindseysbrown/Auto-eD/tags it looks empty.
