### 2022-03-21 lindseysbrown/Auto-eD#28 @rasbt
Installation instructions
Moving to the next point in the JOSE reviewer instructions, there is the item

> Installation: Does installation proceed as outlined in the documentation? (and documentation is sufficient?)

Where would I find the installation instructions? I think it's a bit tricky because it's a webapp. I am not 100% sure how the instructions should look like and maybe @labarba can help clarify whether this should be

a. Instructions for how to deploy the code here as a webapp on Heroku?
b. How to install the underlying Python package?
