### 2022-03-21 lindseysbrown/Auto-eD#27 @lindseysbrown
I didn't realize the tags didn't push automatically.  (I could see it locally with git describe).  Pushed now.
