### 2022-03-22 lindseysbrown/Auto-eD#28 @lindseysbrown
For the package, we point users to DeveloperDocumentation.ipynb (both in the README and the paper).  That includes a set of installation instructions.  (Was this unclear or the question is really about how to evaluate installation instructions of a web app?)
