### 2022-03-22 lindseysbrown/Auto-eD#28 @rasbt
Thanks, now that you mention it, I see there is a link in the paper. However, I don't think the repo README.md (https://github.com/lindseysbrown/Auto-eD) contains a link or reference. Could you add it?

PS: It is currently a bit unclear if this (the instruction for the package installation) is enough, or if installation instructions also should contain something on setting up the webserver @labarba. I guess the challenge here is that there are really two types of things: the python package and the GUI &  webserver, and which one do the reviewer instructions / checklists refer to?
