### 2022-03-22 lindseysbrown/Auto-eD#28 @lindseysbrown
README.md now explicitly refers readers to DeveloperDocumentation.ipynb for installation instructions for the package.

The web application is accessible via the URL so there's not really any installation to be done.
