### 2022-03-22 lindseysbrown/Auto-eD#28 @rasbt
Thanks for adding it. Yeah, since the software can also be run locally via `python ADapp.py`, we probably don't need to review the webapp part.
