### 2022-03-22 #161 @labarba
👋 hi! When we started JOSE, we planned to have [two submission types](https://openjournals.readthedocs.io/en/jose/submitting.html): software papers, and papers about learning modules (teaching content of some size bigger than a lecture and smaller than a course). When submitting, authors choose one article type, and according to that, our editorial system posts the correct review checklist. On a couple of occasions, we've had submissions that bridge the two article types, but we don't have an editorial process for these edge cases.

I request that you use the review checklist for the software (as that is the submission type chosen by the authors) but comment here on your assessment of the other relevant parts of this project.
