### 2022-04-19 #161 @rasbt
Hey @labarba , I have yet another question about the review process 😅. When I reviewed papers for JOSS and JOSE in the past, as far as I remember, there have always been unit tests (and even a CI) as it is best practice for open source.

In the reviewer guidelines it says

>  Tests: Are there automated tests or manual steps described so that the function of the software can be verified?

So, it does sound like unit tests (or a CI setup) are not required. Technically, one could walk through the examples in ([DeveloperDocumentation.ipynb](https://github.com/lindseysbrown/Auto-eD/blob/master/DeveloperDocumentation.ipynb)) and check that everything is correct and then rely on jupyter notebook file-diffs to check whether the results can be reproduced on one's own machine. However, this would be quite a burden on the reviewer/user/developer, and in the spirit of open source best-practices, I am wondering if it is okay to suggest to the package maintainers to add unit tests to test the software and verify the functionality on one's own machine?
