### 2022-04-19 #161 @labarba
Our submission guidelines are not exhaustive, because they cannot be, and we do rely on community standards. We should make reference to those community standards when our review guidelines are insufficient.

In regards to testing, the key is that to be eligible for publication, the software needs to adhere to minimum standards that enable reusability: the code should be immediately useful to others. For JOSE, the software should be useful in teaching and learning settings. If the learners do not need to look at or modify the code, as the tool is used more like an app, any tests and documentation target potential contributors, who need to have a way to check correctness. This may exclude unit tests, but at least should have some ability to do system or regression testing. Relying on Jupyter-notebook diffs is surely inadequate for that?
