### 2022-04-19 #161 @rasbt
It's sometimes tricky to strike the balance between being a critical reviewer and trying to improve the submission, but not coming across as asking for unreasonable things.

Thanks for the response on that. I do think that in this case unit tests not only help me as a reviewer, but it could also help users who want to run it locally and use their package verify that everything works as intended in their environment. And, of course it makes contributing and updating this package in the future easier.

I think in this case it doesn't have to be a comprehensive new suite of unit tests hooked up to a CI. A reasonable balance between extra work and utility would be to convert the current examples in the developer doc into unit tests so that someone can run them from their computer and verify the actual results match the expected results.
