### 2022-04-19 lindseysbrown/Auto-eD#32 @rasbt
Unit tests to verify functionality
Sorry to push more things onto your todo list, but in the context of the JOSE review, it would be really helpful to have some basic unit tests in place.

Let me try to be reasonable and not ask for setting up automatic tests for the UI :). What I have in mind is basically "just" copying the examples and outputs you have in DeveloperDocumentation.ipynb into a unit test setup that works with pytest or nose.

Adding a CI service via a GitHub workflow would of course be great, but this is up to you. As a reviewer, it is sufficient for me to just run the code locally on my computer and knowing that the actual results match the expected results. In order to do that right now I would have to run the code in the DeveloperDocumentation.ipynb and check the results with my local results cell by cell. This is a bit tedious, to be honest. And also thinking about potential users who want to use the library and app locally, this would be tedious as well.

So, in the spirit of open source, there should be a unit test suite set up so that I just need to run a command like

    pytest ./ADlib

I am really not expecting you to develop new unit tests. For me, just converting the examples you have in DeveloperDocumentation.ipynb  would already be sufficient.
