### 2022-04-25 lindseysbrown/Auto-eD#32 @lindseysbrown
Hi @rasbt thank you for emphasizing the utility of testing for developers of the code (our main intention when writing was the use of the web app to complement classroom instruction - which the user wouldn't need to run tests for - perhaps we should have selected learning module instead of software submission...).  In the initial development, we did have some unit tests written so we have added those back to the repo in addition to adding tests that explicitly match the examples in the documentation.  Running pytest now works as you describe (and we have noted this option under installation in the documentation notebook).
