### 2022-05-14 lindseysbrown/Auto-eD#32 @rasbt
Awesome, thanks.  Just toyed with the unit tests via pytest and everything works.

Also, the small note you added should satisfy the "described" aspect in the reviewer checklist in my opinion :)

> Tests: Are there automated tests or manual steps described so that the function of the software can be verified?
