### 2022-05-14 lindseysbrown/Auto-eD#34 @rasbt
Missing community guidelines
Hey, another thing for the JOSE review is the following point:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Maybe you could just add it as a separate sections in the Readme. Maybe two sections

## Getting Help

...

## Contributing

...

should suffice?
