### 2022-05-18 lindseysbrown/Auto-eD#34 @lindseysbrown
Hi @rasbt, we had noted in the DeveloperDocumentation under Future Work how to contribute to the software, but I have now added the notes you requested in the README as well.
