### 2022-05-19 lindseysbrown/Auto-eD#35 @rasbt
Add missing references
Hey,

almost there. Just a few final things to check off for the review! There is the item

> References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

I just see that the paper's Reference section is blank. Could you fill it in with the required details?
