### 2022-05-20 lindseysbrown/Auto-eD#35 @lindseysbrown
Hi @rasbt , I think that's how the journal's paper building system is set up.  (If you check the repos for recently published papers, the References section shows blank for all of them in github.)  References are listed in paper.bib.  I just rechecked doi's - not finding one for Greiwank 1989 but I have been able to track them down for the other citations.
