### 2022-05-20 lindseysbrown/Auto-eD#35 @rasbt
Sounds good! Could you add the correct citation for the Andreas Griewank entry though please:

```
@article{griewank1989automatic,
  title={On automatic differentiation},
  author={Griewank, Andreas},
  journal={Mathematical Programming: recent developments and applications},
  volume={6},
  number={6},
  pages={83--107},
  year={1989},
  publisher={Amsterdam}
}
```
this doesn't look right to me. I think it's a book, and the publisher is Kluwer. So, this article would be a book chapter part of the book.
