### 2022-05-20 #161 @rasbt
The authors just addressed the last items on my checklist, and everything looks good to me now from my end.
