### 2022-05-22 #161 @labarba
👋  @Atcold — You have some unchecked items in the review checklist for this JOSE submission. Can you give us an update on your review? Have you been able to complete a full review of the submission? Are there any changes you recommended to the authors that remain pending? Thank you! 🙏
