### 2022-06-10 #161 @labarba
Hi @Atcold — we don't have the ability to compile a diff. The changes are of course tracked by commits on the target repository: https://github.com/lindseysbrown/Auto-eD/commits/master

@lindseysbrown — could you summarize the changes, to facilitate the review?
