### 2022-06-10 #161 @lindseysbrown
Hi @Atcold, the main clarifying change in the paper is shown in this commit to outline the three different levels.  I think this should address your main concern about the lack of clarity in how someone should interact with the tool/learning module/software here:
https://github.com/lindseysbrown/Auto-eD/commit/2d32cb2120ee9d471d25c3966a77d5ea501fedd8

Other changes in response to reviewer comments are on the issues tab (adding versioning, tests, and some edits to a citation).
