### 2022-07-04 #161 @labarba
Oops. I still get the order of things mixed up. A few steps needed:

1. Please give the article proof one last read-through to find any remaining typos, and approve the proof for publication.
2. Make a tagged release on the target repository, and report the version number here.
3. Make an archive of the repository on Zenodo or an equivalent service, and report the archive DOI here.
