### 2022-07-05 #161 @lindseysbrown
@labarba My coauthors and I have reviewed the proof and made a couple minor edits to correct typos.

I have created a tagged release: v1.0.2 (https://github.com/lindseysbrown/Auto-eD/tree/v1.0.2)

And archived that release on Zenodo: 10.5281/zenodo.6800009

Please let me know if you need any further information.
