### 2022-07-05 #161 @labarba
Interesting. The DOI itself does not resolve at http://doi.org/10.5555/3122009.3242010
