### 2022-07-05 #161 @labarba
No need. The release and archive are really targeting the package. The paper will be deposited by JOSE with its own DOI and will be archived too.
