### 2023-11-22 #161 @arfon
@labarba – the submitting author contacted me (on admin@theoj.org) to ask for an update to their paper (see this change: https://github.com/lindseysbrown/Auto-eD/commit/7c1eef58e5d568c739c4902b7b3706ae7a22bc92). This should be possible with a simple `@editorialbot reaccept` but before I do this I wanted to get an explicit :+1: from you.
